ADD: mission1-4 + mock data feed 구현
