+++ public/data/yujulee/mockdata.json
@@ -1,0 +1,6 @@
+[
+  { "id": 1, "link": "/images/img.yujulee/photo8.jpeg", "text": "yuzu-like" },
+  { "id": 2, "link": "/images/img.yujulee/photo2.jpeg", "text": "wannagohome" },
+  { "id": 3, "link": "/images/img.yujulee/photo6.jpg", "text": "mockdata" },
+  { "id": 4, "link": "/images/img.yujulee/photo7.jpg", "text": "simsimhae" }
+]
