+++ src/pages/yujulee/Main/Mainyuju.js
@@ -1,5 +1,6 @@
-import React, { useState } from 'react';
+import React from 'react';
 import './Mainyuju.scss';
+import Feed from './Feed';
 
 //props로 중복코드제거
 function Story({ src, figcaption }) {
@@ -23,27 +24,6 @@ function Recommend({ src, id, friends }) {
 }
 
 function Mainyuju() {
-  //게시버튼 색바꾸기
-  const [blue, setBlue] = useState(false);
-
-  function turnBlue() {
-    return input.length > 0 ? setBlue(true) : setBlue(false);
-  }
-  //댓글 구현
-  const [input, setInput] = useState('');
-  const [commentArr, setCommentArr] = useState([]);
-
-  function saveComment(event) {
-    setInput(event.target.value);
-  }
-  function commentArray() {
-    setCommentArr([...commentArr, input]);
-    setInput('');
-  }
-  function SingleComment({ content }) {
-    return <div>yuzu_like {content}</div>;
-  }
-
   return (
     <>
       <nav className="navbar">
@@ -73,57 +53,7 @@ function Mainyuju() {
               figcaption="soojinee"
             />
           </section>
-
-          <article className="feedarticle">
-            <div className="feedname">
-              <img
-                className="preview"
-                src="/images/img.yujulee/photo6.jpg"
-                width="100%"
-                alt="story"
-              />
-              <span>yuzu_like</span>
-              <span className="friends">대한민국</span>
-              <i className="fa-solid fa-ellipsis" />
-            </div>
-            <img
-              className="feedimg"
-              src="/images/img.yujulee/photo8.jpeg"
-              width="100%"
-              alt="story"
-            />
-            <br />
-            <div className="feedtalk">
-              <i className="fa-regular fa-heart press" />
-              <i className="fa-regular fa-comment" />
-              <i className="fa-regular fa-paper-plane" />
-              <br />
-              <div>좋아요 25개</div>
-              <div> yuzu_like 놀러가구싶다!!</div>
-              <div className="date">1일 전</div>
-              {commentArr.map((comments, index) => (
-                <SingleComment key={index} content={comments} />
-              ))}
-            </div>
-            <div className="comment">
-              <i className="smileface fa-regular fa-face-smile" />
-              <input
-                type="text"
-                className="commentbox"
-                placeholder="댓글 달기..."
-                onKeyUp={turnBlue}
-                onChange={saveComment}
-                value={input}
-              />
-              <button
-                className={blue ? 'aftercomment' : 'beforecomment'}
-                disabled={blue ? false : true}
-                onClick={commentArray}
-              >
-                게시
-              </button>
-            </div>
-          </article>
+          <Feed />
         </main>
         <aside className="sidebar">
           <div className="feedname">
@@ -163,6 +93,10 @@ function Mainyuju() {
             id="i_likeeee"
             friends="회원님을 위한 추천"
           />
+          <div className="asidearr">
+            소개 도움말 홍보 센터 API 채용 정보 개인정보처리방침 <br />
+            약관 위치 인기 계정 해시태그 언어
+          </div>
         </aside>
       </div>
     </>
