+++ src/pages/yujulee/Main/Mainyuju.scss
@@ -175,3 +175,8 @@ ul {
     margin-top: 7px;
   }
 }
+.asidearr {
+  color: #8e8e8e;
+  font-size: 12px;
+  margin-left: 12px;
+}
